v0.2.2 - add CLI argument parsing and improve excluded list
I've added argument parsing with Thor (`vaccine-spotter help` will show 
all options) and redid the way already-checked/excluded locations are 
dealt with — the program will now check when appointments at a location 
become *unavailable* and will then remove its ID from the 'excluded' 
blacklist (instead of clearing the list when the current time was 
divisible by 5, as was previously the case). It will also print a 
message that appoitnments are now unavailable at a location.